SmartMed/medlabs/system/serializers.py
# Python imports

# Django imports

# 3rd party imports
from rest_framework import serializers

# App imports
from .models import Patient, MedicalRecord

class PatientSerializer(serializers.ModelSerializer):
    class Meta:
        model = Patient
        fields= '__all__'

class MedicalRecordSerializer(serializers.ModelSerializer):
    class Meta:
        model = MedicalRecord
        fields = '__all__'
